Nodes - Cleaned up Color Gradient nodes
