+++ src/Artemis.VisualScripting/Nodes/Color/ColorGradientNode.cs
@@ -1,19 +1,20 @@
 ﻿using Artemis.Core;
-using Artemis.Core.Events;
 using Artemis.VisualScripting.Nodes.Color.Screens;
 using SkiaSharp;
 using System.Collections.Specialized;
 
 namespace Artemis.VisualScripting.Nodes.Color;
 
-[Node("Color Gradient", "Outputs a configurable Color Gradient", "Color", OutputType = typeof(ColorGradient))]
+[Node("Color Gradient (Simple)", "Outputs a color gradient with the given colors", "Color", OutputType = typeof(ColorGradient))]
 public class ColorGradientNode : Node<ColorGradient, ColorGradientNodeCustomViewModel>
 {
     #region Constructors
 
     public ColorGradientNode()
-        : base("Color Gradient", "Outputs a configurable Color Gradient")
+        : base("Color Gradient", "Outputs a color gradient with the given colors")
     {
+        Storage = ColorGradient.GetUnicornBarf();
+
         _inputPins = new();
 
         Output = CreateOutputPin<ColorGradient>();
@@ -97,6 +98,12 @@ public class ColorGradientNode : Node<ColorGradient, ColorGradientNodeCustomView
             RemovePin(pin);
             _inputPins.Remove(pin);
         }
+        
+        int index = 0;
+        foreach (var item in _inputPins)
+        {
+            item.Name = $"Color #{++index}";
+        }
     }
 
     #endregion
@@ -108,101 +115,4 @@ public class ColorGradientNode : Node<ColorGradient, ColorGradientNodeCustomView
     public OutputPin<ColorGradient> Output { get; }
 
     #endregion
-}
-
-[Node("Color Gradient", "Outputs a Color Gradient from colors and positions", OutputType = typeof(ColorGradient))]
-public class ColorGradientFromPinsNode : Node
-{
-    public OutputPin<ColorGradient> Gradient { get; set; }
-    public InputPinCollection<SKColor> Colors { get; set; }
-    public InputPinCollection<Numeric> Positions { get; set; }
-
-    public ColorGradientFromPinsNode() : base("Color Gradient", "Outputs a Color Gradient from colors and positions")
-    {
-        Colors = CreateInputPinCollection<SKColor>("Colors", 0);
-        Positions = CreateInputPinCollection<Numeric>("Positions", 0);
-        Gradient = CreateOutputPin<ColorGradient>("Gradient");
-
-        Colors.PinAdded += OnPinAdded;
-        Colors.PinRemoved += OnPinRemoved;
-        Positions.PinAdded += OnPinAdded;
-        Positions.PinRemoved += OnPinRemoved;
-    }
-
-    private void OnPinRemoved(object? sender, SingleValueEventArgs<IPin> e)
-    {
-        var colorsCount = Colors.Count();
-        var positionsCount = Positions.Count();
-        if (colorsCount == positionsCount)
-            return;
-
-        while (colorsCount > positionsCount)
-        {
-            var pinToRemove = Colors.Last();
-            Colors.Remove(pinToRemove);
-
-            --colorsCount;
-        }
-
-        while (positionsCount > colorsCount)
-        {
-            var pinToRemove = Positions.Last();
-            Positions.Remove(pinToRemove);
-            --positionsCount;
-        }
-
-        RenamePins();
-    }
-
-    private void OnPinAdded(object? sender, SingleValueEventArgs<IPin> e)
-    {
-        var colorsCount = Colors.Count();
-        var positionsCount = Positions.Count();
-        if (colorsCount == positionsCount)
-            return;
-
-        while (colorsCount < positionsCount)
-        {
-            Colors.Add(Colors.CreatePin());
-
-            ++colorsCount;
-        }
-
-        while (positionsCount < colorsCount)
-        {
-            Positions.Add(Positions.CreatePin());
-
-            ++positionsCount;
-        }
-
-        RenamePins();
-    }
-
-    private void RenamePins()
-    {
-        int colors = 0;
-        foreach (var item in Colors)
-        {
-            item.Name = $"Color #{++colors}";
-        }
-
-        int positions = 0;
-        foreach(var item in Positions)
-        {
-            item.Name = $"Position #{++positions}";
-        }
-    }
-
-    public override void Evaluate()
-    {
-        var stops = new List<ColorGradientStop>();
-        var colors = Colors.Pins.ToArray();
-        var positions = Positions.Pins.ToArray();
-        for (int i = 0; i < colors.Length; i++)
-        {
-            stops.Add(new ColorGradientStop(colors[i].Value, positions[i].Value));
-        }
-
-        Gradient.Value = new ColorGradient(stops);
-    }
 }
